Make local template package compilable
The only way to extend the original text/template package is to make a local copy and ensure that it refers to itself when importing text/template.

I also had to make a local copy of internal/fmtsort since it is not possible to refer to internal packages from outside golang.